Fixed encoding bug. I has assumed base32 rather than base64
The default in the code is base32 encoding, but when I made the key.txt file I used an option to specifically choose base64 encoding. Which is why I got the mysterious error message:

     Error illegal base32 data at input byte 0

Of course it was base32 illegal at byte 0! It wasn't base32 at all. It was base64. Duh!